requirements.txt
playsound==1.2.1
pyobjc-core==3.2.1
pyobjc-framework-Cocoa==3.2.1
pyobjc-framework-Quartz==3.2.1
